requirements.txt
psycopg2-binary==2.9.10   # For PostgreSQL access
nextcord                  # Discord bot framework
python-dotenv             # For loading .env configs
